spotify-recommendation/requirements.txt
Flask==2.1.3
lyricsgenius==3.0.1
numpy==1.23.1
python-dotenv==1.0.0
scikit_learn==1.2.2
spotipy==2.23.0
